third_party/blink/web_tests/external/wpt/permissions-policy/payment-extension-allowed-by-permissions-policy-attribute.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] WebAuthn "payment" extension can be enabled in cross-origin iframe using allow="payment" attribute
  assert_equals: undefined. expected "OK" but got "NotSupportedError"
Harness: the test ran to completion.
